구조물 청사진 설치 안되던거 해결, Keymgr 클릭하자마자 해제 안되도록, Hover 최적화

+++ 숙제/3개월차/API 시연회/RimWorld/RimWorld/BoltActionRifle.cpp
@@ -65,7 +65,7 @@ void CBoltActionRifle::Render(HDC hDC)
     int		iScrollX = (int)CScrollMgr::Get_Instance()->Get_ScrollX();
     int		iScrollY = (int)CScrollMgr::Get_Instance()->Get_ScrollY();
 
-    POINT tPoint = CCamera::Get_Instance()->WorldToScreen(m_tRect.left, m_tRect.top);
+    POINT tPoint = CCamera::Get_Instance()->WorldToScreen((float)m_tRect.left, (float)m_tRect.top);
     float fZoom = CCamera::Get_Instance()->Get_Zoom();
     Graphics Grapics(hDC);
 
@@ -87,6 +87,6 @@ void CBoltActionRifle::Render(HDC hDC)
     Grapics.TranslateTransform((REAL)-centerX, (REAL)-centerY);
 
     // 이미지 그리기
-    Grapics.DrawImage(m_pImage, tPoint.x + iScrollX, tPoint.y + iScrollY, 64 * fZoom, 64 * fZoom);
+    Grapics.DrawImage(m_pImage, tPoint.x + iScrollX, tPoint.y + iScrollY, int(64 * fZoom), int(64 * fZoom));
 
 }
